backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openfeature-sdk
openfeature-provider-flagd
launchdarkly-server-sdk
python-dotenv